requirements.txt
streamlit
pillow
dask
dask[dataframe]
polars
sqlite3
modin
koalas
pandas
sqlalchemy
vaex
ray